Update OmniLogger tests to fix SimpleTest memory issues.
SimpleTest chokes performing an 'is equal' comparison on
the Omnilog_entry class, now that said class contains a
reference to the EE global object.

The reason is that SimpleTest converts the objects to
arrays, and compares each key in turn. The reference
to the EE object therefore becomes an enormous nested
array, and SimpleTest can't cope.